+++ crates/node/core/Cargo.toml
@@ -18,7 +18,7 @@ reth-cli.workspace = true
 reth-cli-util.workspace = true
 reth-fs-util.workspace = true
 reth-db = { workspace = true, features = ["mdbx"] }
-reth-db-api.workspace = true
+reth-db-models.workspace = true
 reth-storage-errors.workspace = true
 reth-provider.workspace = true
 reth-network = { workspace = true, features = ["serde"] }
