+++ crates/storage/db-api/src/models/blocks.rs
@@ -1,7 +1,7 @@
 //! Block related models and types.
 
 use reth_codecs::{add_arbitrary_tests, Compact};
-use reth_primitives::{Header, Withdrawals, B256};
+use reth_primitives::{Header, B256};
 use serde::{Deserialize, Serialize};
 
 /// The storage representation of a block's ommers.
@@ -13,15 +13,6 @@ use serde::{Deserialize, Serialize};
 pub struct StoredBlockOmmers {
     /// The block headers of this block's uncles.
     pub ommers: Vec<Header>,
-}
-
-/// The storage representation of block withdrawals.
-#[derive(Debug, Default, Eq, PartialEq, Clone, Serialize, Deserialize, Compact)]
-#[cfg_attr(any(test, feature = "arbitrary"), derive(arbitrary::Arbitrary))]
-#[add_arbitrary_tests(compact)]
-pub struct StoredBlockWithdrawals {
-    /// The block withdrawals.
-    pub withdrawals: Withdrawals,
 }
 
 /// Hash of the block header.
