+++ crates/storage/db-models/src/lib.rs
@@ -6,4 +6,8 @@ pub use accounts::AccountBeforeTx;
 
 /// Blocks
 pub mod blocks;
-pub use blocks::StoredBlockBodyIndices;
+pub use blocks::{StoredBlockBodyIndices, StoredBlockWithdrawals};
+
+/// Client Version
+pub mod client_version;
+pub use client_version::ClientVersion;
