+++ examples/rpc-db/Cargo.toml
@@ -11,7 +11,7 @@ jsonrpsee.workspace = true
 reth.workspace = true
 reth-chainspec.workspace = true
 reth-db.workspace = true
-reth-db-api.workspace = true
+reth-db-models.workspace = true
 reth-node-ethereum.workspace = true
 reth-provider = { workspace = true, features = ["test-utils"] }
 tokio = { workspace = true, features = ["full"] }
